レビュー指摘を反映: CLI 分離 + stop() の dispose 削除
1. CLI エントリポイントを src/cli.ts に分離し、daemon.ts のインポート時に
   TranscriptWatcher やシグナルハンドラが起動する副作用を解消
2. stop() から speaker.dispose() を削除し、flushText() でエンキューした
   メッセージが破棄される問題を解消
