+++ src/daemon.ts
@@ -78,7 +78,6 @@ export class Daemon {
     this.debounceTimers.clear();
 
     await this.watcher.close();
-    this.speaker?.dispose();
   }
 
   /**
@@ -213,28 +212,3 @@ export function formatToolUse(
   }
 }
 
-// -- CLI entry point --
-
-async function main(): Promise<void> {
-  const daemon = new Daemon();
-
-  const shutdown = (): void => {
-    process.stderr.write("[cc-voice-reporter] shutting down...\n");
-    void daemon.stop().then(() => {
-      process.exit(0);
-    });
-  };
-
-  process.on("SIGINT", shutdown);
-  process.on("SIGTERM", shutdown);
-
-  await daemon.start();
-  process.stderr.write("[cc-voice-reporter] daemon started\n");
-}
-
-main().catch((error: unknown) => {
-  process.stderr.write(
-    `[cc-voice-reporter] fatal: ${error instanceof Error ? error.message : String(error)}\n`,
-  );
-  process.exit(1);
-});
